# 8339335: set number of parallel jobs when building webkit