@kevinrushforth 2024-09-03
I'll run this through or CI as well as do some local builds. The one concern I have is that `NUM_COMPILE_THREADS` is set to 1 on macOS, meaning that this will increase the build time on macOS.

/reviewers 2