@johanvos 2024-09-03
It is easy to create (yet) another property for this, but that will make the already complicated build.gradle even more complicated, so if possible I'd like to avoid that.

I read the cryptic comment in build.gradle
`// On Mac we limit it to 1 by default due to problems running gcc in parallel`
What is the problem that is being referred to here? And why would that problem occur in the other native compile jobs but not in the (resource intensive) webkit compile jobs?