@johanvos 2024-09-03
I see 3 options:
1. the current approach in this PR. Drawback: huge increase in build time on mac
2. a separate option for the number of parallel jobs for webkit. Drawbacks: not elegant because
* we don't know why there are no issues compiling webkit with parallel jobs while there are issues compiling JavaFX native code with parallel jobs.
* adding another property increases complexity
3. Finding and fixing the cause for the parallel compilation issues on mac (JDK-8089417). Drawback: finding the cause may take a very long time -- if it is even still possible to reproduce
4. Removing the 1-thread restriction for mac. Drawback: unless we do (3), we don't know if this is not going to cause issues.

I'm ok with either approach. I consider option 3 the most elegant one, and option 2 the most pragmatic one (but also the least elegant).