@kevinrushforth 2024-09-03
As you say, option 3 is the cleanest, however I believe it will very quickly morph into option 4. I am reasonably certain that [JDK-8089417](https://bugs.openjdk.org/browse/JDK-8089417) can be closed as "Cannot reproduce" because:

* The font T2K code that was hitting the failure is no longer part of our repo; T2K was a closed-source font rasterizer that we removed in JDK 10 via [JDK-8187147](https://bugs.openjdk.org/browse/JDK-8187147).
* Xcode switched from gcc to clang by default several years ago
 
I will locally remove the workaround, and then run the following in a loop a few hundred times to check whether we see any issues:

```
$ gradle cleanNative sdk
```

I did it a few times on my MacBook M1 and the only impact I see of removing the workaround is that a rebuild of the SDK without media or WebKit (so just rebuilding the graphics / glass native code) goes from 27 seconds to 12 seconds. :)