@johanvos 2024-09-03
> Presuming that my testing shows no problems, I propose the following:
> 
> 1. I will close [JDK-8089417](https://bugs.openjdk.org/browse/JDK-8089417) as "Cannot reproduce"
> 2. I will file a new bug to remove the workaround that currently disabled parallel compilation on mac for the graphics module (I think that's cleaner from a bookkeeping perspective than repurposing JDK-8089417)
> 3. You can add the removal of the logic that forced `NUM_COMPILE_THREADS` to 1 on macOS to this PR, and `/issue add` the new JBS issue
> 
> Thoughts?

Sounds like a plan. I'll add that to the PR once you confirm testing went as expected.