@kevinrushforth 2024-09-03
Btw, I ran another CI build with the `NUM_COMPILE_THREADS=1` workaround for Mac removed, and the build times are good again. In fact, the x64 build runs even faster than before -- possibly because WebKit build defaults to a lower number than num_processors, and it happens to be 12 on that x64 build system.

Here are the updated numbers:

OS / Arch|Current|This PR|PR + don't limit num threads on macOS
---------|-------|-------|-------
macOS / aarch64|13m:41s|56m:36s|13m:43s
macOS / x64|28m:30s|2h:24m:52s|23m:01s

I should also add that the Linux and Windows build times are largely unaffected by this PR (which is good). They do run slightly faster with your patch than current, but the percentage gain is small.