@kevinrushforth 2024-09-03
> Sounds like a plan. I'll add that to the PR once you confirm testing went as expected.

So far so good. I'll let it run a b it longer, but already I'm over 120 iterations with no failures.