@kevinrushforth 2024-09-03
I'm done testing and have closed [JDK-8089417](https://bugs.openjdk.org/browse/JDK-8089417) as "Cannot reproduce". I filed a new bug:

[JDK-8339505](https://bugs.openjdk.org/browse/JDK-8339505): Enable parallel compilation of native code on macOS

Go ahead and remove the force-NUM_COMPILE_THREADS -to-1 workaround and add that new issue to this PR.