@johanvos 2024-09-03
/issue add JDK-8339505