@andy-goryachev-oracle 2024-09-03
I am using this command to build:

```
(rm -rf ./build ; gradle -PCOMPILE_MEDIA=true -PCOMPILE_WEBKIT=true all 2>&1) | tee ~/`date +"build-%Y-%m%d-%H%M%S"`.log
```

on macOS 14.6.1 / M1 Pro

looking at the Activity Monitor, I see 1 thread per clang compiler. Is this expected, or am I supposed to see more concurrent threads?

![Screenshot 2024-09-03 at 13 25 55](https://github.com/user-attachments/assets/96085195-d71d-4ff7-9d49-7090efc7a892)