@kevinrushforth 2024-09-03
This setting (nor likely any setting we can, or would want to, control) doesn't affect the parallelism of an individual clang process, but rather controls the maximum number of concurrent clang processes that will be launched.

The next effect of this patch on macOS should be that we will now do parallel compilation for native code in the graphics module by default. There should be no change, by default, in the parallelism of WebKit compilation (there would have been before the most recent commit that Johan did).

What this patch will do is to apply the same `NUM_COMPILE_THREADS` flag to native WebKit as well as native graphics compilation. Prior to this fix, setting `NUM_COMPILE_THREADS` to, say 2, would have been used for the graphics build and ignored for the WebKit build.