@andy-goryachev-oracle 2024-09-03
```
-PNUM_COMPILE_THREADS=1: BUILD SUCCESSFUL in 1h 9m 44s
-PNUM_COMPILE_THREADS=10: BUILD SUCCESSFUL in 17m 4s
-PNUM_COMPILE_THREADS=32: BUILD SUCCESSFUL in 19m 49s
```